requirements.txt
mysql-connector-python
tablib
tablib[xls]
openpyxl
